reworked loss function, tests pass
