+++ blackbirds/infer/__init__.py
@@ -1,0 +1,3 @@
+from .smd import SMD
+from .vi import VI
+from .mcmc import MALA, MCMC
