+++ blackbirds/models/random_walk.py
@@ -40,8 +40,8 @@ class RandomWalk(Model):
         - `params`: a tensor of shape (1,) containing the logit probability of moving forward at each timestep.
         - `x`: a tensor of shape (n,) containing the time-series of positions.
 
-        !!! danger 
-            probability is given in logit, so the input is transformed using the sigmoid function. 
+        !!! danger
+            probability is given in logit, so the input is transformed using the sigmoid function.
         """
         p = torch.sigmoid(params)
         logits = torch.vstack((p, 1 - p)).log()
