--- test/test_example.py
@@ -1,2 +1,0 @@
-def test__simple():
-    assert 1 == 1
